requirements.txt
Django
django-model-utils
PyYAML